requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-js-asset==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
Werkzeug
django-countries
django-extensions
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.0
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-js-asset==3.0.1
django-payments==3.0.1
django-phonenumber-field==8.0.0
django-seo==0.3.5
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
idna==3.10
django-jazzmin
kombu==5.4.2
MarkupSafe==3.0.2
phonenumberslite==8.13.53
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise
gunicorn
crispy-bootstrap4
crispy-bootstrap5
django-crispy-forms
django-widget-tweaks
